Blood Castle / Devil Square, combate, mundo interactivo e inventario

+++ src/Render/BMD_LegacyDraw.cpp
@@ -199,7 +199,7 @@ void __cdecl FUN_004414d0(void *model, char a, int b, float frame, int flags,
                     // IDA: v56 = (__int64)WorldTime + 931 * v33;  (v33 = indice
                     // de vertice). El port ponia `frame` — el indice de MALLA —
                     // donde va WorldTime, asi que la onda quedaba congelada.
-                    int iVar13 = (int)DAT_05826e08 + 0x3a3 * iVar7;
+                    int iVar13 = (long long)DAT_05826e08 + 0x3a3 * iVar7;
                     float sinVal = (float)fsin((double)iVar13 * (double)_DAT_005528c4);
                     float *pfVar12 = (float *)((char*)&DAT_0584621c + ((int)frame * 15000 + iVar7) * 3 * 4);
                     int normBase = ((int)*psVar15 + (int)frame * 15000) * 0xc;
